greyscale converter

+++ src/main/scala/Models/Pixel/GreyScalePixel.scala
@@ -1,5 +1,5 @@
 package Models.Pixel
 
-case class GreyScalePixel (private val value: Int) extends Pixel {
-  def getValue: Int = value
+case class GreyScalePixel (private val value: Double) extends Pixel {
+  def getValue: Double = value
 }
